Add support for SMD-only footprints, and for test pins.
Test pins are broken out as TP1-TP6; however, as per the datasheet,
TP4/TP5 are ignored.

The main reason to break out test pins is access to the USB DP/DM pins
(if you want to break out a Pico to an external USB connector), and
access to the BOOTSEL pin.